pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "code-lm"
version = "0.3.0"
authors = [
  { name="Panagiotis897", email="orion256business@gmail.com" }
]
description = "An AI coding assistant using various LLM models."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "requests>=2.25.0",
    "PyYAML>=6.0",
    "tiktoken>=0.6.0",
    "questionary>=2.0.0",
]

[project.urls]
"Homepage" = "https://github.com/Panagiotis897/lm-code"
"Bug Tracker" = "https://github.com/Panagiotis897/lm-code/issues"

[project.scripts]
lmcode = "lm_code.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
